+++ Classes/Domain/Model/Coordinate.php
@@ -1,9 +1,4 @@
 <?php
-/**
- * Date: 04/2020
- *
- * @author Jessica Schlierenkamp <mail@schlierenkamp.de>
- */
 
 declare(strict_types = 1);