+++ Classes/Domain/Model/Filter/ProjectFilter.php
@@ -1,24 +1,13 @@
 <?php
-/**
- * Date: 04/2020
- *
- * @author Jessica Schlierenkamp <mail@schlierenkamp.de>
- */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace JS\Wechange\Domain\Model\Filter;
 
-class ProjectFilter implements FilterInterface
+class ProjectFilter extends AbstractSortableFilter
 {
-    public const ORDER_ASC = 'ASC';
-    public const ORDER_DESC = 'DESC';
-
     private int $parent;
     private string $tag;
-    private int $limit;
-    private string $orderBy;
-    private string $orderDir;
 
     /**
      * @codeCoverageIgnore
@@ -27,63 +16,28 @@ class ProjectFilter implements FilterInterface
         int $parent = 0,
         string $tag = '',
         int $limit = 10,
+        int $offset = 0,
         string $orderBy = '',
         string $orderDir = self::ORDER_ASC
     ) {
+        parent::__construct($limit, $offset, $orderBy, $orderDir);
+
         $this->parent = $parent;
         $this->tag = $tag;
-        $this->limit = $limit;
-        $this->orderBy = $orderBy;
-        $this->orderDir = $orderDir;
     }
 
-    /**
-     * @return int
-     */
     public function getParent(): int
     {
         return $this->parent;
     }
 
-    /**
-     * @return string
-     */
     public function getTag(): string
     {
         return $this->tag;
     }
 
-    /**
-     * @return int
-     */
-    public function getLimit(): int
-    {
-        return $this->limit;
-    }
-
-    /**
-     * @return string
-     */
-    public function getOrderBy(): string
-    {
-        return $this->orderBy;
-    }
-
-    /**
-     * @return string
-     */
-    public function getOrderDir(): string
-    {
-        return $this->orderDir;
-    }
-
-    /**
-     * @return string
-     */
     public function buildQueryString(): string
     {
-        $queryString = [];
-
         if ($parent = $this->getParent()) {
             $queryString[] = 'parent=' . $parent;
         }
@@ -96,6 +50,10 @@ class ProjectFilter implements FilterInterface
             $queryString[] = 'limit=' . $limit;
         }
 
-        return implode('&', $queryString);
+        if ($offset = $this->getOffset()) {
+            $queryString[] = 'offset=' . $offset;
+        }
+
+        return implode('&', $queryString ?? []);
     }
 }