+++ Classes/Domain/Model/Filter/StatisticFilter.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JS\Wechange\Domain\Model\Filter;
+
+use JS\Wechange\Domain\Model\Coordinate;
+
+class StatisticFilter extends AbstractFilter
+{
+    public function buildQueryString(): string
+    {
+        return '';
+    }
+}