+++ Classes/Domain/Repository/EventRepository.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace JS\Wechange\Domain\Repository;
+
+/**
+ * @codeCoverageIgnore
+ */
+class EventRepository extends AbstractSortableRepository
+{
+    public function getApiSlug(): string
+    {
+        return 'events/';
+    }
+}